Update CreateRelease.ps1
Deploy Package steps need the packagereferencename to be blank, the way it was written here will always pull the package id which will break deploy package steps. This instead looks at the deployment process and pulls in the information it needs and allows you to add any logic necessary for versioning.

This also changes it so the version of the release is the next increment, but I'm not sure if you want to keep that.