Create AddTenantTagToTenant.ps1
Should this go here, or the tenanttag folder? my gut says here since we're making an API call to tenants endpoint.

Line 10 and 11 is a bit jank-looking but I was having issues getting the return I wanted so I cannibalized some other script in our repo and it worked so I kept it.