Create DownloadFromAzurePushToOctopusServer
This code downloads all packages from a UniversalPackagesFeed in Azure Devops.

After downloading the packages a new release is created, with the same name as the version of the packages that was found in azure.

It expects that all packages in that feed has the same version. Because that's how it was set up in our organization.

The powershell script expects the Azure CLI to be installed. If it's not it will download and start installation of it. This will create a pop-up.

There's place for a lot of improvements. And if your usecase is not excactly as ours, it may be a fine starting-point